Remove node factory
